requirements.txt
streamlit==1.32.0
openai==1.78.1
python-dotenv==1.0.0 